requirements.txt
flask==3.1.0
requests==2.26.0
pythion-dotenv==1.0.0